feat(oruga, oruga-next): Export individual components
It allows components to be registered locally instead of globally (see [1]).
Some IDE can automatically insert these imports which greatly reduces the hassle
of using tree shaking.

It doesn't handle CSS imports yet, but the old implementation did not either.
A promising solution is to use a bundler plugin, which is adopted by Element
Plus [2].

[1]: https://vuejs.org/guide/components/registration.html#local-registration
[2]: https://github.com/element-plus/unplugin-element-plus