fix(autocomplete): close after selection
When an autocomplete value is selected via the autocomplete's
`setSelected()` method, the autocomplete's v-model value will usually be
updated, and the dropdown will usually scheduled to be closed on the
next tick. However, updating the v-model value triggers the v-model
watcher, which would then schedule the dropdown to be open again on the
next tick.

This change updates the v-model watcher to not do anything when the new
value is the same as the currently selected value.