fix(autocomplete): select head/foot exclusively
When the `selectable-header` (or `selectable-footer` prop) is true, it
was possible to use the up & down keys to turn on the visual "hovered"
state for both the header (or footer) option and another regular option
at the same time.

This change centralizes the logic of setting the `hoveredOption`,
`headerHovered`, and `footerHovered` refs in the `setHovered()`
function, ensuring that only one can be set at a time.