fix(autocomplete): apply aria combox pattern
The ARIA attributes used by the autocomplete menu and options
represented the menu as if it were a standalone modal dialog.

The change applies the Combobox Pattern from the ARIA Authoring
Practices Guide to the autocomplete input, menu, and options -- which
should allow assistive technologies to recognize the menu as a list of
options available for the input, and to identify the currently "hovered"
option.

See https://www.w3.org/WAI/ARIA/apg/patterns/combobox/ for details.