fix(autocomplete): clear itemRefs on re-render
Using the up & down keys to change the "hovered" state triggers a
re-render of the autocomplete menu options -- but not of the
autocomplete component itself. The autocomplete's own `onBeforeUpdate`
handler is not called when this happens; and so using it to clear the
`itemRefs` array resulted in the previous set of options not being
removed.

The main manifestation of this bug was that it was not possible to use
the down key to access a selectable-footer option.

This change clears the `itemRefs` array whenever the `setItemRef()`
function is called for the first menu option; now the `itemRefs` array
will be cleared both whenever the full autocomplete component is
re-rendered, and whenever just the individual options are re-rendered.