fix(test): remove extra copy of Vitest
This is the result of removing and re-adding Vitest:

```
npm remove -w @oruga-ui/oruga-next \
  vitest \
  @vitest/browser \
  @vitest/coverage-istanbul \
  vitest-browser-vue
npm i -w @oruga-ui/oruga-next --save-dev \
  vitest@^3.1.0 \
  @vitest/browser@^3.1.0 \
  @vitest/coverage-istanbul@^3.1.0 \
  vitest-browser-vue@^0.2.0
```

I'm not exactly sure why the old version was sticking around, but it may
have had something to do with a peer dependency of `vitest-browser-vue`.
The old version doesn't seem to break CI, but it does break typechecking
with the Vue language server.